splitted router integration tests to multiple files
